script/newsletter/pyproject.toml
[project]
name = "my-newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "html2text>=2024.2.26",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.11.0",
    "transformers>=4.52.4",
    "python-dateutil>=2.9.0.post0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
